python_requirements.txt
torch
torchvision
transformers
Pillow
numpy
scikit-learn
matplotlib
tqdm
omegaconf
tqdm
PyYAML
einops
lightning
taming-transformers-rom1504
hydra-core
pandas
wandb
pytorch_fid
h5py
tensorboard
pot
tbparse